+++ .next/static/chunks/webpack.js
@@ -192,7 +192,7 @@
 /******/ 	
 /******/ 	/* webpack/runtime/getFullHash */
 /******/ 	!function() {
-/******/ 		__webpack_require__.h = function() { return "5b6759c0d1930b37"; }
+/******/ 		__webpack_require__.h = function() { return "9896e53903b5adf9"; }
 /******/ 	}();
 /******/ 	
 /******/ 	/* webpack/runtime/global */